Setup moderator service.
This service  checks for content of a newly created comment. If said comment
includes any inappropriate word; then that comment is marked as rejected; else
as approved.

This new service highlights micorservices where the data might need some time
to be processed. But where we still respond immediately to users with
some response. In this case, their comment is marked as
`pending-review`.